Remove Elasticsearch REST API
CrateDB uses it's own engine to write and read data. Accessing or
even writing data using the ES API could lead to inconsitent
data/state.
Also lot of functionality exposed by this API was never offically
supported by CrateDB.
