+++ blackbox/docs/src/doc_tests/tests.py
@@ -417,7 +417,6 @@ def create_doctest_suite():
         settings={
             'license.enterprise': 'true',
             'lang.js.enabled': 'true',
-            'es.api.enabled': 'true',
             'psql.port': GLOBAL_PORT_POOL.get(),
         }
     )
