+++ blob/src/test/java/io/crate/integrationtests/BlobIntegrationTestBase.java
@@ -45,7 +45,6 @@ import java.util.Map;
 import java.util.function.Consumer;
 import java.util.stream.Stream;
 
-import static io.crate.rest.CrateRestMainAction.ES_API_ENABLED_SETTING;
 import static org.elasticsearch.common.network.NetworkModule.HTTP_ENABLED;
 import static org.elasticsearch.http.HttpTransportSettings.SETTING_CORS_ALLOW_ORIGIN;
 import static org.elasticsearch.http.HttpTransportSettings.SETTING_CORS_ENABLED;
@@ -71,7 +70,6 @@ public abstract class BlobIntegrationTestBase extends ESIntegTestCase {
         return Settings.builder()
             .put(super.nodeSettings(nodeOrdinal))
             .put(HTTP_ENABLED.getKey(), true)
-            .put(ES_API_ENABLED_SETTING.getKey(), true)
             .put(SETTING_CORS_ENABLED.getKey(), true)
             .put(SETTING_CORS_ALLOW_ORIGIN.getKey(), "*")
             .put(SETTING_HTTP_COMPRESSION.getKey(), false)
