--- es/es-server/src/main/java/org/elasticsearch/rest/action/search/RestClearScrollAction.java
@@ -1,74 +1,0 @@
-/*
- * Licensed to Elasticsearch under one or more contributor
- * license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright
- * ownership. Elasticsearch licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.elasticsearch.rest.action.search;
-
-import org.elasticsearch.action.search.ClearScrollRequest;
-import org.elasticsearch.client.node.NodeClient;
-import org.elasticsearch.common.Strings;
-import org.elasticsearch.common.settings.Settings;
-import org.elasticsearch.rest.BaseRestHandler;
-import org.elasticsearch.rest.RestController;
-import org.elasticsearch.rest.RestRequest;
-import org.elasticsearch.rest.action.RestStatusToXContentListener;
-
-import java.io.IOException;
-import java.util.Arrays;
-
-import static org.elasticsearch.rest.RestRequest.Method.DELETE;
-
-public class RestClearScrollAction extends BaseRestHandler {
-    public RestClearScrollAction(Settings settings, RestController controller) {
-        super(settings);
-
-        controller.registerHandler(DELETE, "/_search/scroll", this);
-        controller.registerHandler(DELETE, "/_search/scroll/{scroll_id}", this);
-    }
-
-    @Override
-    public String getName() {
-        return "clear_scroll_action";
-    }
-
-    @Override
-    public RestChannelConsumer prepareRequest(final RestRequest request, final NodeClient client) throws IOException {
-        String scrollIds = request.param("scroll_id");
-        ClearScrollRequest clearRequest = new ClearScrollRequest();
-        clearRequest.setScrollIds(Arrays.asList(splitScrollIds(scrollIds)));
-        request.withContentOrSourceParamParserOrNull((xContentParser -> {
-            if (xContentParser != null) {
-                // NOTE: if rest request with xcontent body has request parameters, values parsed from request body have the precedence
-                try {
-                    clearRequest.fromXContent(xContentParser);
-                } catch (IOException e) {
-                    throw new IllegalArgumentException("Failed to parse request body", e);
-                }
-            }
-        }));
-
-        return channel -> client.clearScroll(clearRequest, new RestStatusToXContentListener<>(channel));
-    }
-
-    private static String[] splitScrollIds(String scrollIds) {
-        if (scrollIds == null) {
-            return Strings.EMPTY_ARRAY;
-        }
-        return Strings.splitStringByCommaToArray(scrollIds);
-    }
-}
