--- es/es-server/src/main/java/org/elasticsearch/rest/action/search/RestSearchScrollAction.java
@@ -1,73 +1,0 @@
-/*
- * Licensed to Elasticsearch under one or more contributor
- * license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright
- * ownership. Elasticsearch licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.elasticsearch.rest.action.search;
-
-import org.elasticsearch.action.search.SearchScrollRequest;
-import org.elasticsearch.client.node.NodeClient;
-import org.elasticsearch.common.settings.Settings;
-import org.elasticsearch.rest.BaseRestHandler;
-import org.elasticsearch.rest.RestController;
-import org.elasticsearch.rest.RestRequest;
-import org.elasticsearch.rest.action.RestStatusToXContentListener;
-import org.elasticsearch.search.Scroll;
-
-import java.io.IOException;
-
-import static org.elasticsearch.common.unit.TimeValue.parseTimeValue;
-import static org.elasticsearch.rest.RestRequest.Method.GET;
-import static org.elasticsearch.rest.RestRequest.Method.POST;
-
-public class RestSearchScrollAction extends BaseRestHandler {
-    public RestSearchScrollAction(Settings settings, RestController controller) {
-        super(settings);
-
-        controller.registerHandler(GET, "/_search/scroll", this);
-        controller.registerHandler(POST, "/_search/scroll", this);
-        controller.registerHandler(GET, "/_search/scroll/{scroll_id}", this);
-        controller.registerHandler(POST, "/_search/scroll/{scroll_id}", this);
-    }
-
-    @Override
-    public String getName() {
-        return "search_scroll_action";
-    }
-
-    @Override
-    public RestChannelConsumer prepareRequest(final RestRequest request, final NodeClient client) throws IOException {
-        String scrollId = request.param("scroll_id");
-        SearchScrollRequest searchScrollRequest = new SearchScrollRequest();
-        searchScrollRequest.scrollId(scrollId);
-        String scroll = request.param("scroll");
-        if (scroll != null) {
-            searchScrollRequest.scroll(new Scroll(parseTimeValue(scroll, null, "scroll")));
-        }
-
-        request.withContentOrSourceParamParserOrNull(xContentParser -> {
-            if (xContentParser != null) {
-                // NOTE: if rest request with xcontent body has request parameters, values parsed from request body have the precedence
-                try {
-                    searchScrollRequest.fromXContent(xContentParser);
-                } catch (IOException e) {
-                    throw new IllegalArgumentException("Failed to parse request body", e);
-                }
-            }});
-        return channel -> client.searchScroll(searchScrollRequest, new RestStatusToXContentListener<>(channel));
-    }
-}
